wip: prepare for docker
